% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_stan.R
\name{age_gender_class_stan}
\alias{age_gender_class_stan}
\title{Age gender and Class Bayesian logistic regression with Stan}
\usage{
age_gender_class_stan(
  age,
  pclass,
  sex_cats,
  sex_idx,
  survived,
  test_sex_idx,
  test_age,
  test_pclass,
  ...
)
}
\arguments{
\item{age}{Numeric vector of age values.}

\item{pclass}{integer value of the passengers class status.}

\item{sex_cats}{integer value of total number of sexes possible (in this case it's 2).}

\item{sex_idx}{integer value identifying the passenger sex.}

\item{survived}{integer vector of output values, this is binary.}

\item{test_sex_idx}{integer value identifying the passenger sex from the test dataset.}

\item{test_age}{Numeric vector of age values from test dataset.}

\item{test_pclass}{integer value of the test dataset for passenger class status.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Age gender and Class Bayesian logistic regression with Stan
}
