% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcEBErG.R
\name{hosp}
\alias{hosp}
\title{All hospitalizations and hospitalizations with respiratory diagnosis.}
\usage{
hosp(start, end)
}
\arguments{
\item{start}{start season (year started).}

\item{end}{end season (year started).}
}
\value{
Return season, agegrp, hospilitazions with respiratory diagnosis, all hospitalilazions, mean length of resp-hospitalitazions from season \code{start} to \code{end}.
}
\description{
Demand ODBC access to IB_EpiLPR
}
\examples{
\dontrun{
hosp(2010, 2017)
}
}
