% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{impute_passenger_age}
\alias{impute_passenger_age}
\title{Impute Passenger Age}
\usage{
impute_passenger_age(data)
}
\arguments{
\item{data}{a dataframe of passenger details that include age.}
}
\description{
Fill in null age values with some estimate of the passengers
age.
}
