% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcEBErG.R
\name{influ}
\alias{influ}
\title{Hospitalizations with a positive influenza sample during hospitalizaion.}
\usage{
influ(start, end)
}
\arguments{
\item{start}{start season (year started).}

\item{end}{end season (year started).}
}
\value{
season, agegrp, hospilitazions with a positive influenza specimen, mean length of these hospitalizations from season \code{start} to \code{end}
}
\description{
Hospitalizations with a positive influenza sample during hospitalizaion.
}
