% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcEBErG.R
\name{pop}
\alias{pop}
\title{Population by age group by 1st January.}
\usage{
pop(start, end)
}
\arguments{
\item{start}{start year.}

\item{end}{end year.}
}
\value{
Return season, agegrp, population by January 1 from season \code{start} to \code{end}.
}
\description{
Population by age group by 1st January.
}
\examples{
\dontrun{
pop(2010, 2017)
}
}
