% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcEBErG.R
\name{PosPct}
\alias{PosPct}
\title{Positive percentage among all specimens.}
\usage{
PosPct(start, end)
}
\arguments{
\item{start}{start season (year started).}

\item{end}{end season (year started).}
}
\value{
Return season, agegrp, number of positive specimen and all specimens, and positive percentage from season \code{start} to \code{end}.
}
\description{
Positive percentage among all specimens.
}
\examples{
\dontrun{
PosPct(2010, 2017)
}
}
