% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_stan.R
\name{stan-models}
\alias{stan-models}
\alias{age_gender_stan}
\alias{age_gender_class_stan}
\title{Interface to Stan Models}
\usage{
age_gender_stan(age, sex_cats, sex_idx, survived, test_sex_idx, test_age, ...)

age_gender_class_stan(
  age,
  class_1,
  class_2,
  class_3,
  sex_cats,
  sex_idx,
  survived,
  test_sex_idx,
  test_age,
  test_class_1,
  test_class_2,
  test_class_3,
  ...
)
}
\arguments{
\item{age}{Numeric vector of age values.}

\item{sex_cats}{integer value of total number of sexes possible (in this case it's 2).}

\item{sex_idx}{integer value identifying the passenger sex.}

\item{survived}{integer vector of output values, this is binary.}

\item{test_sex_idx}{integer value identifying the passenger sex from the test dataset.}

\item{test_age}{Numeric vector of age values from test dataset.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{pclass}{integer value of the passengers class status.}

\item{test_pclass}{integer value of the test dataset for passenger class status.}
}
\description{
These functions are the interface to the logistic regression models fit using stan.
}
\section{Function Descriptions}{

\describe{
\item{\code{age_gender_stan()}}{
Age gender Bayesian logistic regression with Stan
}
\item{\code{age_gender_class_stan()}}{
Age gender and Class Bayesian logistic regression with Stan
}
}
}

