% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_test}
\alias{titanic_test}
\title{Titanic Passenger Testing Dataset}
\format{
A data frame with 418 rows and 11 variables:
\describe{
\item{PassengerId}{}
\item{Pclass}{}
\item{Name}{}
\item{Sex}{}
\item{Age}{}
\item{SibSp}{}
\item{Parch}{}
\item{Ticket}{}
\item{Fare}{}
\item{Cabin}{}
\item{Embarked}{}
...
}
}
\source{
\url{https://www.kaggle.com/c/titanic/data}
}
\usage{
titanic_test
}
\description{
A dataset containing information for a subset of the
passengers on the titanic. Used for testing the accuracy
of the predictive model.
}
\keyword{datasets}
