% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_train}
\alias{titanic_train}
\title{Titanic Passenger Training Dataset}
\format{
A data frame with 891 rows and 12 variables:
\describe{
\item{PassengerId}{}
\item{Survived}{}
\item{Pclass}{}
\item{Name}{}
\item{Sex}{}
\item{Age}{}
\item{SibSp}{}
\item{Parch}{}
\item{Ticket}{}
\item{Fare}{}
\item{Cabin}{}
\item{Embarked}
...
}
}
\source{
\url{https://www.kaggle.com/c/titanic/data}
}
\usage{
titanic_train
}
\description{
A dataset containing information for a subset of the
passengers on the titanic. Used for training a predictive
model.
}
\keyword{datasets}
