% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic-package.R
\docType{package}
\name{titanic-package}
\alias{titanic-package}
\alias{titanic}
\title{The 'titanic' package.}
\description{
Encapsulate the work used in the Titanic Kaggle Challenge.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
